refactor directory, deleted BRCode management
