+++ src/main/java/it/gov/pagopa/payment/controller/payment/CommonPaymentControllerImpl.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payment.controller;
+package it.gov.pagopa.payment.controller.payment;
 
 import it.gov.pagopa.common.performancelogger.PerformanceLog;
 import it.gov.pagopa.payment.dto.qrcode.TransactionCreationRequest;
