+++ src/test/java/it/gov/pagopa/payment/controller/payment/BasePaymentControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payment.controller;
+package it.gov.pagopa.payment.controller.payment;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
