+++ src/test/java/it/gov/pagopa/payment/service/payment/common/CommonCreationServiceImplTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.payment.service.payment.qrcode;
+package it.gov.pagopa.payment.service.payment.common;
 
 import com.mongodb.client.result.UpdateResult;
 import it.gov.pagopa.common.web.exception.ClientException;
@@ -16,6 +16,8 @@ import it.gov.pagopa.payment.model.RewardRule;
 import it.gov.pagopa.payment.model.TransactionInProgress;
 import it.gov.pagopa.payment.repository.RewardRuleRepository;
 import it.gov.pagopa.payment.repository.TransactionInProgressRepository;
+import it.gov.pagopa.payment.service.payment.qrcode.QRCodeCreationService;
+import it.gov.pagopa.payment.service.payment.qrcode.QRCodeCreationServiceImpl;
 import it.gov.pagopa.payment.test.fakers.MerchantDetailDTOFaker;
 import it.gov.pagopa.payment.test.fakers.TransactionCreationRequestFaker;
 import it.gov.pagopa.payment.test.fakers.TransactionInProgressFaker;
